+++ packages/template-retail-react-app/package.json
@@ -11,9 +11,9 @@
     "analyze-build": "cross-env MOBIFY_ANALYZE=true npm run build",
     "build": "npm run build-translations && pwa-kit-dev build",
     "build-translations": "npm run extract-default-translations && npm run compile-translations",
-    "compile-translations": "formatjs compile-folder --ast app/translations app/translations/compiled",
-    "compile-translations:pseudo": "formatjs compile --ast app/translations/en-US.json --out-file app/translations/compiled/en-XB.json --pseudo-locale en-XB",
-    "extract-default-translations": "node ./scripts/build-translations.js",
+    "compile-translations": "formatjs compile-folder --ast translations translations/compiled",
+    "compile-translations:pseudo": "formatjs compile --ast translations/en-US.json --out-file translations/compiled/en-XB.json --pseudo-locale en-XB",
+    "extract-default-translations": "node ./scripts/extract-default-messages.js",
     "format": "pwa-kit-dev format '**/*.{js,jsx}'",
     "lint": "pwa-kit-dev lint '**/*.{js,jsx}'",
     "lint:fix": "npm run lint -- --fix",
